Add iitial functionality for proportional ratio of text size to its scale
