+++ Messenger/Pages/Chats/SemiComponents/HelperComponents/Footer/ReplyAndEditMenu.tsx
@@ -50,10 +50,15 @@ class ReplyAndEditMenu extends Component<ReplyAndEditMenuProps> {
             <View style={styles.dataContainer}>
               <View>
                 <Text style={styles.usernameText}>{displayName}</Text>
-                <Text style={styles.messageText}>{
+                <Text 
+                  style={styles.messageText}
+                  numberOfLines={1}
+                >
+                  {
                   isReply?(replyMessage?.content!.length>DEFAULT_CHARS_PER_LINE?text.slice(0,DEFAULT_CHARS_PER_LINE)+'...':replyMessage?.content):
-                  (editMessage?.content!.length>DEFAULT_CHARS_PER_LINE?text.slice(0,DEFAULT_CHARS_PER_LINE)+'...':editMessage?.content)
-                }</Text>
+                    (editMessage?.content!.length>DEFAULT_CHARS_PER_LINE?text.slice(0,DEFAULT_CHARS_PER_LINE)+'...':editMessage?.content)
+                  }
+                </Text>
               </View>
               <TouchableOpacity 
                 onPress={cancelReplyAndEdit}
