+++ Messenger/Pages/Chats/SemiComponents/HelperComponents/Footer/Styles/CenterPartOfFooter.ts
@@ -1,5 +1,5 @@
 import { StyleSheet } from "react-native";
-import { height, width } from "../../../ChatConstants";
+import { FOOTER_INNER_TEXTINPUT_GAP, MAX_FOOTER_HEIGHT, getCustomFontSize, height, width } from "../../../ChatConstants";
 
 
 export const styles = StyleSheet.create({
@@ -12,8 +12,9 @@ export const styles = StyleSheet.create({
     backgroundColor: 'rgba(255, 255, 255, 0.7)',
     color: 'rgb(137, 130, 130)',
     paddingHorizontal: 15,
+    fontSize: getCustomFontSize(15),
     width: '60%',
     zIndex: 10,
-    maxHeight: height * 0.12
+    maxHeight: MAX_FOOTER_HEIGHT - FOOTER_INNER_TEXTINPUT_GAP
   },
 })
