+++ Messenger/Pages/Chats/SemiComponents/HelperComponents/Header/Styles/PinnedMessageView.ts
@@ -1,11 +1,10 @@
 import { StyleSheet } from "react-native";
-import { screenHeight, screenWidth } from "../../../../../ChatList/Constants/ConstantsForChatlist";
-import { SOFT_MENU_BAR_HEIGHT } from "../../../ChatConstants";
+import { height, width } from "../../../ChatConstants";
 
 export const styles = StyleSheet.create({
   mainContainer: { 
     position: 'absolute', 
-    bottom: -screenHeight*0.185 + SOFT_MENU_BAR_HEIGHT*0.6, 
+    bottom: -height*0.15, 
     backgroundColor: '#fff', 
     overflow: 'hidden', 
     borderRadius: 9999, 
@@ -18,13 +17,13 @@ export const styles = StyleSheet.create({
     position: "absolute",
     left: 0,
     right: 0,
-    height: screenHeight,
-    width: screenWidth,
+    height: height,
+    width: width,
   },
   container: { 
     flexDirection: 'row', 
     justifyContent: 'space-between', 
-    width: screenWidth*0.9, 
+    width: width*0.9, 
     paddingVertical: 10, 
     paddingHorizontal: 20, 
     alignItems: 'center'
