+++ Messenger/Pages/Chats/SemiComponents/Interfaces/IMessageItem.tsx
@@ -1,21 +1,20 @@
 
 import ILastWatchedMessage from "../../../../dao/Models/Chats/ILastWatchedMessage";
 import User from "../../../../dao/Models/User";
-import { coordY } from "../Screens/PinnedMessageScreen";
 import { Layout } from "./GeneralInterfaces/ILayout";
 import { MessageProps } from "./GeneralInterfaces/IMessage";
-
+import { coordY } from "./IPinnedMessageScreen";
 
 export interface MessageItemProps { 
   item: MessageProps;
   listOfMessages: MessageProps[];
   setMessageMenuVisible: {(arg0: Layout, arg1: boolean):void}; 
   flatListRef?: React.MutableRefObject<null>;
-  coordsY: [number[]] | coordY[];
+  coordsY?: coordY[];
+  // setCoordsY: (newCoordsY: any) => void;
   author: User;
   users: User[];
   messageID: number;
-  setCoordsY: (newCoordsY: any) => void;
   userMessageLastWatched?: ILastWatchedMessage | undefined;
   selecting: boolean;
   pinnedMessageHandler?: (message: number, coord: number) => void;
