+++ Messenger/Pages/Chats/SemiComponents/SVG/HeaderBranchButton.tsx
@@ -1,19 +1,29 @@
 import * as React from "react";
 import Svg, { Path } from "react-native-svg";
-import { screenHeight, screenWidth } from "../../../ChatList/Constants/ConstantsForChatlist";
+import { screenHeight } from "../../../ChatList/Constants/ConstantsForChatlist";
 
 function HeaderBranchButton() {
   return (
     <Svg
-      width={screenWidth * 0.17} 
-      height={screenHeight * 0.044}
-      viewBox="0 0 12 13"
+      width={screenHeight * 0.035} 
+      height={screenHeight * 0.035}
+      viewBox="0 0 23 24"
       fill="none"
     >
       <Path
-        d="M2.143 2.64c.17 0 .334-.064.454-.178a.594.594 0 00.189-.43.594.594 0 00-.189-.432.661.661 0 00-.454-.178.661.661 0 00-.455.178.594.594 0 00-.188.431c0 .162.068.317.188.431.12.114.284.179.455.179zm2.143-.609c0 .833-.528 1.55-1.286 1.861v2.23a3.554 3.554 0 011.714-.434h2.572C8.23 5.688 9 4.957 9 4.063v-.17C8.242 3.58 7.714 2.863 7.714 2.03 7.714.91 8.674 0 9.857 0 11.041 0 12 .909 12 2.031c0 .833-.528 1.55-1.286 1.861v.17c0 1.796-1.535 3.25-3.428 3.25H4.714C3.77 7.313 3 8.043 3 8.938v.17c.758.313 1.286 1.029 1.286 1.862 0 1.122-.96 2.031-2.143 2.031C.959 13 0 12.091 0 10.969c0-.833.528-1.55 1.286-1.861V3.892C.528 3.58 0 2.864 0 2.032 0 .908.959 0 2.143 0c1.184 0 2.143.909 2.143 2.031zm6.214 0a.594.594 0 00-.188-.43.661.661 0 00-.455-.18.661.661 0 00-.454.18.594.594 0 00-.189.43c0 .162.068.317.189.431.12.114.284.179.454.179.17 0 .334-.065.455-.179a.594.594 0 00.188-.43zm-8.357 9.547c.17 0 .334-.064.454-.178a.593.593 0 00.189-.431.593.593 0 00-.189-.431.661.661 0 00-.454-.179.661.661 0 00-.455.179.593.593 0 00-.188.43c0 .162.068.317.188.432.12.114.284.178.455.178z"
-        fill="#2B1D1D"
-        strokeWidth={0.3}
+        d="M7.705 4.824c0 1.628-1.308 2.936-2.908 2.936-1.6 0-2.909-1.308-2.909-2.936s1.309-2.937 2.909-2.937c1.6 0 2.908 1.309 2.908 2.937z"
+        stroke="#222"
+        strokeWidth={1.8}
+      />
+      <Path
+        d="M4.799 8.02v8.952M4.799 16.973c1.27-9.591 13.329.64 13.964-9.591"
+        stroke="#222"
+        strokeWidth={2}
+      />
+      <Path
+        d="M21.67 4.824c0 1.628-1.309 2.936-2.909 2.936-1.6 0-2.908-1.308-2.908-2.936s1.309-2.937 2.908-2.937c1.6 0 2.909 1.309 2.909 2.937zM7.705 20.17c0 1.629-1.308 2.937-2.908 2.937-1.6 0-2.909-1.308-2.909-2.936 0-1.629 1.309-2.937 2.909-2.937 1.6 0 2.908 1.308 2.908 2.936z"
+        stroke="#222"
+        strokeWidth={1.8}
       />
     </Svg>
   );
